requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.2